lua: move lua output code to generic lua file
So that other Lua scripts (detect) can also start using it.
