+++ src/Makefile.am
@@ -227,7 +227,6 @@ output-json-http.c output-json-http.h \
 output-json-ssh.c output-json-ssh.h \
 output-json-tls.c output-json-tls.h \
 output-lua.c output-lua.h \
-output-lua-common.c output-lua-common.h \
 output-packet.c output-packet.h \
 output-streaming.c output-streaming.h \
 output-tx.c output-tx.h \
@@ -320,6 +319,7 @@ util-ip.h util-ip.c \
 util-logopenfile.h util-logopenfile.c \
 util-logopenfile-tile.h util-logopenfile-tile.c \
 util-lua.c util-lua.h \
+util-lua-common.c util-lua-common.h \
 util-lua-http.c util-lua-http.h \
 util-magic.c util-magic.h \
 util-memcmp.c util-memcmp.h \
