+++ src/util-lua-common.h
@@ -21,8 +21,8 @@
  * \author Victor Julien <victor@inliniac.net>
  */
 
-#ifndef __OUTPUT_LUA_COMMON_H__
-#define __OUTPUT_LUA_COMMON_H__
+#ifndef __UTIL_LUA_COMMON_H__
+#define __UTIL_LUA_COMMON_H__
 
 #ifdef HAVE_LUA
 
@@ -37,4 +37,4 @@ int LuaRegisterFunctions(lua_State *luastate);
 
 #endif /* HAVE_LUA */
 
-#endif /* __OUTPUT_LUA_COMMON_H__ */
+#endif /* __UTIL_LUA_COMMON_H__ */
